feat(basic-info): extract BasicInfoFields and reuse in UI
Extract shared BasicInfoFields component for name/description inputs and progress. Update BasicInfoCard and BasicInfoStep to consume it. Files: basic-info-fields.tsx, basic-info-card.tsx, basic-info-step.tsx.
