+++ apps/web/src/components/tenant/basic-info-fields.tsx
@@ -1,0 +1,148 @@
+"use client";
+
+import React, { useMemo } from "react";
+import { Input } from "@/components/ui/input";
+import { Label } from "@/components/ui/label";
+import { Textarea } from "@/components/ui/textarea";
+
+type Props = {
+  nameValue: string;
+  descriptionValue: string;
+  onChange: (
+    e: React.ChangeEvent<HTMLInputElement | HTMLTextAreaElement>
+  ) => void;
+};
+
+export function BasicInfoFields({
+  nameValue,
+  descriptionValue,
+  onChange,
+}: Props) {
+  const nameProgress = useMemo(() => {
+    const length = nameValue?.length || 0;
+    return Math.min((length / 100) * 100, 100);
+  }, [nameValue]);
+
+  const descriptionProgress = useMemo(() => {
+    const length = descriptionValue?.length || 0;
+    return Math.min((length / 500) * 100, 100);
+  }, [descriptionValue]);
+
+  return (
+    <div className="space-y-6">
+      <div className="space-y-4">
+        <div className="flex items-center gap-3">
+          <div>
+            <Label htmlFor="name" className="font-semibold text-foreground">
+              Property Name
+            </Label>
+          </div>
+        </div>
+
+        <div className="space-y-3">
+          <Input
+            id="name"
+            name="name"
+            value={nameValue}
+            onChange={onChange}
+            placeholder="e.g., Luxury Beachfront Villa with Ocean Views"
+            className="h-12 text-base border-2 border-slate-200 focus:ring-2 focus-visible:border-primary/50 focus-visible:ring-primary/10 transition-all duration-200"
+            maxLength={100}
+            required
+          />
+
+          <div className="space-y-2">
+            <div className="flex justify-between items-center">
+              <span className="text-xs text-muted-foreground">
+                {nameValue?.length || 0} of 100 characters
+              </span>
+              <span
+                className={`text-xs font-medium ${
+                  nameProgress > 90
+                    ? "text-amber-600"
+                    : nameProgress > 80
+                    ? "text-orange-600"
+                    : "text-green-600"
+                }`}
+              >
+                {nameProgress.toFixed(0)}%
+              </span>
+            </div>
+            <div className="h-2 w-full bg-slate-100 rounded-full overflow-hidden">
+              <div
+                className={`h-full transition-all duration-300 rounded-full ${
+                  nameProgress > 90
+                    ? "bg-gradient-to-r from-amber-400 to-orange-500"
+                    : nameProgress > 80
+                    ? "bg-gradient-to-r from-orange-400 to-red-500"
+                    : "bg-gradient-to-r from-green-400 to-emerald-500"
+                }`}
+                style={{ width: `${nameProgress}%` }}
+              />
+            </div>
+          </div>
+        </div>
+      </div>
+
+      <div className="space-y-2">
+        <div className="flex items-center">
+          <div>
+            <Label
+              htmlFor="description"
+              className="font-semibold text-foreground"
+            >
+              Property Description
+            </Label>
+          </div>
+        </div>
+
+        <div className="space-y-3">
+          <Textarea
+            id="description"
+            name="description"
+            value={descriptionValue}
+            onChange={onChange}
+            placeholder="Tell guests about your property's unique features, location highlights, amenities, and what makes it special..."
+            rows={5}
+            maxLength={500}
+            className="flex min-h-[120px] w-full rounded-lg border-2 border-slate-200 bg-background px-4 py-3 text-base ring-offset-background placeholder:text-muted-foreground focus-visible:outline-none focus-visible:border-primary/50 focus-visible:ring-2 focus-visible:ring-primary/10 transition-all duration-200 resize-none"
+            required
+          />
+
+          <div className="space-y-2">
+            <div className="flex justify-between items-center">
+              <span className="text-xs text-muted-foreground">
+                {descriptionValue?.length || 0} of 500 characters
+              </span>
+              <span
+                className={`text-xs font-medium ${
+                  descriptionProgress > 90
+                    ? "text-amber-600"
+                    : descriptionProgress > 80
+                    ? "text-orange-600"
+                    : "text-green-600"
+                }`}
+              >
+                {descriptionProgress.toFixed(0)}%
+              </span>
+            </div>
+            <div className="h-2 w-full bg-slate-100 rounded-full overflow-hidden">
+              <div
+                className={`h-full transition-all duration-300 rounded-full ${
+                  descriptionProgress > 90
+                    ? "bg-gradient-to-r from-amber-400 to-orange-500"
+                    : descriptionProgress > 80
+                    ? "bg-gradient-to-r from-orange-400 to-red-500"
+                    : "bg-gradient-to-r from-green-400 to-emerald-500"
+                }`}
+                style={{ width: `${descriptionProgress}%` }}
+              />
+            </div>
+          </div>
+        </div>
+      </div>
+    </div>
+  );
+}
+
+export default BasicInfoFields;
